Set up task to import slate projects
